+++ src/ApolloClient.ts
@@ -68,6 +68,10 @@ import {
   addFragmentsToDocument,
 } from './queries/getFromAST';
 
+import {
+  version,
+} from './version';
+
 /**
  * This type defines a "selector" function that receives state from the Redux store
  * and returns the part of it that is managed by ApolloClient
@@ -108,6 +112,9 @@ export default class ApolloClient {
   public shouldForceFetch: boolean;
   public dataId: IdGetter;
   public fieldWithArgs: (fieldName: string, args?: Object) => string;
+  public version: string;
+
+  private devToolsHookCb: Function;
 
   /**
    * Constructs an instance of {@link ApolloClient}.
@@ -156,6 +163,7 @@ export default class ApolloClient {
     addTypename = true,
     queryTransformer,
     customResolvers,
+    connectToDevTools,
   }: {
     networkInterface?: NetworkInterface,
     reduxRootKey?: string,
@@ -170,6 +178,7 @@ export default class ApolloClient {
     addTypename?: boolean,
     queryTransformer?: any,
     customResolvers?: CustomResolverMap,
+    connectToDevTools?: boolean,
   } = {}) {
     if (reduxRootKey && reduxRootSelector) {
       throw new Error('Both "reduxRootKey" and "reduxRootSelector" are configured, but only one of two is allowed.');
@@ -178,7 +187,7 @@ export default class ApolloClient {
     if (reduxRootKey) {
       console.warn(
           '"reduxRootKey" option is deprecated and might be removed in the upcoming versions, ' +
-          'please use the "reduxRootSelector" instead.'
+          'please use the "reduxRootSelector" instead.',
       );
       this.reduxRootKey = reduxRootKey;
     }
@@ -226,6 +235,22 @@ export default class ApolloClient {
     this.mutate = this.mutate.bind(this);
     this.setStore = this.setStore.bind(this);
     this.resetStore = this.resetStore.bind(this);
+
+    // Attach the client instance to window to let us be found by chrome devtools, but only in
+    // development mode
+    const defaultConnectToDevTools =
+      typeof process === 'undefined' || (process.env && process.env.NODE_ENV !== 'production') &&
+      typeof window !== 'undefined' && (!(window as any).__APOLLO_CLIENT__);
+
+    if (typeof connectToDevTools === 'undefined') {
+      connectToDevTools = defaultConnectToDevTools;
+    }
+
+    if (connectToDevTools) {
+      (window as any).__APOLLO_CLIENT__ = this;
+    }
+
+    this.version = version;
   }
 
   /**
@@ -259,7 +284,7 @@ export default class ApolloClient {
       console.warn(
             '"fragments" option is deprecated and will be removed in the upcoming versions, ' +
             'please refer to the documentation for how to define fragments: ' +
-            'http://dev.apollodata.com/react/fragments.html.'
+            'http://dev.apollodata.com/react/fragments.html.',
       );
       /* istanbul ignore if */
       if (process.env.NODE_ENV !== 'test') {
@@ -309,7 +334,7 @@ export default class ApolloClient {
       console.warn(
             '"fragments" option is deprecated and will be removed in the upcoming versions, ' +
             'please refer to the documentation for how to define fragments: ' +
-            'http://dev.apollodata.com/react/fragments.html.'
+            'http://dev.apollodata.com/react/fragments.html.',
       );
       /* istanbul ignore if */
       if (process.env.NODE_ENV !== 'test') {
@@ -372,7 +397,7 @@ export default class ApolloClient {
       console.warn(
             '"fragments" option is deprecated and will be removed in the upcoming versions, ' +
             'please refer to the documentation for how to define fragments: ' +
-            'http://dev.apollodata.com/react/fragments.html.'
+            'http://dev.apollodata.com/react/fragments.html.',
       );
       /* istanbul ignore if */
       if (process.env.NODE_ENV !== 'test') {
@@ -399,7 +424,7 @@ export default class ApolloClient {
       console.warn(
             '"fragments" option is deprecated and will be removed in the upcoming versions, ' +
             'please refer to the documentation for how to define fragments: ' +
-            'http://dev.apollodata.com/react/fragments.html.'
+            'http://dev.apollodata.com/react/fragments.html.',
       );
       /* istanbul ignore if */
       if (process.env.NODE_ENV !== 'test') {
@@ -427,6 +452,10 @@ export default class ApolloClient {
     return createApolloReducer(this.reducerConfig);
   }
 
+  public __actionHookForDevTools(cb: Function) {
+    this.devToolsHookCb = cb;
+  }
+
   public middleware = () => {
     return (store: ApolloStore) => {
       this.setStore(store);
@@ -434,10 +463,19 @@ export default class ApolloClient {
       return (next: any) => (action: any) => {
         const returnValue = next(action);
         this.queryManager.broadcastNewStore(store.getState());
+
+        if (this.devToolsHookCb) {
+          this.devToolsHookCb({
+            action,
+            state: this.queryManager.getApolloState(),
+            dataWithOptimisticResults: this.queryManager.getDataWithOptimisticResults(),
+          });
+        }
+
         return returnValue;
       };
     };
-  };
+  }
 
   /**
    * This initializes the Redux store that we use as a reactive cache.
@@ -453,7 +491,7 @@ export default class ApolloClient {
           'Cannot initialize the store because "reduxRootSelector" or "reduxRootKey" is provided. ' +
           'They should only be used when the store is created outside of the client. ' +
           'This may lead to unexpected results when querying the store internally. ' +
-          `Please remove that option from ApolloClient constructor.`
+          `Please remove that option from ApolloClient constructor.`,
       );
     }
 
@@ -462,6 +500,19 @@ export default class ApolloClient {
       reduxRootKey: DEFAULT_REDUX_ROOT_KEY,
       initialState: this.initialState,
       config: this.reducerConfig,
+      logger: (store: any) => (next: any) => (action: any) => {
+        const result = next(action);
+
+        if (this.devToolsHookCb) {
+          this.devToolsHookCb({
+            action,
+            state: this.queryManager.getApolloState(),
+            dataWithOptimisticResults: this.queryManager.getDataWithOptimisticResults(),
+          });
+        }
+
+        return result;
+      },
     }));
     // for backcompatibility, ensure that reduxRootKey is set to selector return value
     this.reduxRootKey = DEFAULT_REDUX_ROOT_KEY;
@@ -490,7 +541,7 @@ export default class ApolloClient {
     if (isUndefined(reduxRootSelector(store.getState()))) {
       throw new Error(
           'Existing store does not use apolloReducer. Please make sure the store ' +
-          'is properly configured and "reduxRootSelector" is correctly specified.'
+          'is properly configured and "reduxRootSelector" is correctly specified.',
       );
     }